Move API mutation into ViewDataTable
