+++ src/components/views/ViewDataTable/index.tsx
@@ -1,18 +1,50 @@
 import { FunctionComponent } from 'react';
+import NProgress from 'nprogress';
+import { useRouter } from 'next/router';
+import { useState } from 'react';
 import { Add, Person } from '@material-ui/icons';
 import { Box, Fab } from '@material-ui/core';
 import { DataGridPro, GridColDef } from '@mui/x-data-grid-pro';
+import { useMutation, useQueryClient } from 'react-query';
 
+import postViewColumn from 'fetching/views/postViewColumn';
+import ViewColumnDialog, { ColumnEditorColumnSpec } from 'components/views/ViewColumnDialog';
 import { ZetkinViewColumn, ZetkinViewRow } from 'types/zetkin';
 
 
 interface ViewDataTableProps {
     columns: ZetkinViewColumn[];
-    onAddColumn: () => void;
     rows: ZetkinViewRow[];
+    viewId: string;
 }
 
-const ViewDataTable: FunctionComponent<ViewDataTableProps> = ({ columns, onAddColumn, rows }) => {
+const ViewDataTable: FunctionComponent<ViewDataTableProps> = ({ columns, rows, viewId }) => {
+    const [columnDialogOpen, setColumnDialogOpen] = useState(false);
+    const { orgId } = useRouter().query;
+    const queryClient = useQueryClient();
+
+    const addColumnMutation = useMutation(postViewColumn(orgId as string, viewId), {
+        onError: () => {
+            // TODO: Show error dialog
+            NProgress.done();
+        },
+        onSettled: () => {
+            NProgress.done();
+            queryClient.invalidateQueries(['views', orgId]);
+        },
+        onSuccess: () => queryClient.invalidateQueries(['views', viewId]),
+    });
+
+    const onColumnCancel = () => {
+        setColumnDialogOpen(false);
+    };
+
+    const onColumnSave = (colSpec : ColumnEditorColumnSpec) => {
+        setColumnDialogOpen(false);
+        NProgress.start();
+        addColumnMutation.mutate(colSpec);
+    };
+
     const avatarColumn : GridColDef = {
         disableColumnMenu: true,
         disableExport: true,
@@ -50,7 +82,7 @@ const ViewDataTable: FunctionComponent<ViewDataTableProps> = ({ columns, onAddCo
         renderHeader: () => {
             return (
                 <Box>
-                    <Fab onClick={ () => onAddColumn() } size="small">
+                    <Fab onClick={ () => setColumnDialogOpen(true) } size="small">
                         <Add/>
                     </Fab>
                 </Box>
@@ -77,10 +109,15 @@ const ViewDataTable: FunctionComponent<ViewDataTableProps> = ({ columns, onAddCo
     }));
 
     return (
-        <DataGridPro
-            columns={ gridColumns }
-            rows={ gridRows }
-        />
+        <>
+            <DataGridPro
+                columns={ gridColumns }
+                rows={ gridRows }
+            />
+            { columnDialogOpen && (
+                <ViewColumnDialog onCancel={ onColumnCancel } onSave={ onColumnSave }/>
+            ) }
+        </>
     );
 };
