+++ src/pages/organize/[orgId]/people/views/[viewId].tsx
@@ -1,20 +1,16 @@
 import { GetServerSideProps } from 'next';
 import Head from 'next/head';
-import NProgress from 'nprogress';
-import { useState } from 'react';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
+import { useQuery } from 'react-query';
 
 import getOrg from 'fetching/getOrg';
 import getView from 'fetching/views/getView';
 import getViewColumns from 'fetching/views/getViewColumns';
 import getViewRows from 'fetching/views/getViewRows';
 import { PageWithLayout } from 'types';
-import postViewColumn from 'fetching/views/postViewColumn';
 import { scaffold } from 'utils/next';
 import SingleViewLayout from 'components/layout/organize/SingleViewLayout';
 import ViewDataTable from 'components/views/ViewDataTable';
 import ZetkinQuery from 'components/ZetkinQuery';
-import ViewColumnDialog, { ColumnEditorColumnSpec } from 'components/views/ViewColumnDialog';
 
 
 const scaffoldOptions = {
@@ -61,31 +57,6 @@ type SingleViewPageProps = {
 };
 
 const SingleViewPage: PageWithLayout<SingleViewPageProps> = ({ orgId, viewId }) => {
-    const [columnDialogOpen, setColumnDialogOpen] = useState(false);
-    const queryClient = useQueryClient();
-
-    const addColumnMutation = useMutation(postViewColumn(orgId, viewId), {
-        onError: () => {
-            // TODO: Show error dialog
-            NProgress.done();
-        },
-        onSettled: () => {
-            NProgress.done();
-            queryClient.invalidateQueries(['views', orgId]);
-        },
-        onSuccess: () => queryClient.invalidateQueries(['views', viewId]),
-    });
-
-    const onColumnCancel = () => {
-        setColumnDialogOpen(false);
-    };
-
-    const onColumnSave = (colSpec : ColumnEditorColumnSpec) => {
-        setColumnDialogOpen(false);
-        NProgress.start();
-        addColumnMutation.mutate(colSpec);
-    };
-
     return (
         <ZetkinQuery queries={{
             colsQuery: useQuery(['views', viewId, 'columns'], getViewColumns(orgId, viewId)),
@@ -99,12 +70,9 @@ const SingleViewPage: PageWithLayout<SingleViewPageProps> = ({ orgId, viewId }) 
                     </Head>
                     <ViewDataTable
                         columns={ colsQuery.data }
-                        onAddColumn={ () => setColumnDialogOpen(true) }
                         rows={ rowsQuery.data }
+                        viewId={ viewId }
                     />
-                    { columnDialogOpen && (
-                        <ViewColumnDialog onCancel={ onColumnCancel } onSave={ onColumnSave }/>
-                    ) }
                 </>
             ) }
         </ZetkinQuery>
